requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic-settings==2.2.1
sqlalchemy==2.0.28
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose==3.3.0
python-multipart==0.0.9
email-validator==2.1.1
